refactor: 로또 티켓 관련 패키지 분리
