+++ src/main/kotlin/lotto/domain/WinningResult.kt
@@ -1,5 +1,6 @@
 package lotto.domain
 
+import lotto.domain.lottoticket.LottoTicket
 import java.math.BigDecimal
 import java.math.RoundingMode
