+++ src/main/kotlin/lotto/domain/lottoticket/LottoNumbers.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.lottoticket
 
 class LottoNumbers private constructor(
     val values: Set<LottoNumber>,
